requirements.txt
sklearn
tensorflow
dvc
pandas
numpy
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
spacy
Flask
Flask-Cors

-e .
